+++ examples/radio-app/src/main/java/io/joynr/demo/MyRadioProviderApplication.java
@@ -284,7 +284,6 @@ public class MyRadioProviderApplication extends AbstractJoynrApplication {
             Module backendTransportModules = Modules.EMPTY_MODULE;
 
             if (transport.contains("mqtt")) {
-                joynrConfig.put(MessagingPropertyKeys.PROPERTY_MESSAGING_PRIMARYGLOBALTRANSPORT, "mqtt");
                 backendTransportModules = Modules.combine(backendTransportModules, new HivemqMqttClientModule());
             }
             return Modules.override(runtimeModule).with(backendTransportModules);