+++ examples/stateless-async/stateless-async-car-sim/src/main/java/io/joynr/examples/statelessasync/carsim/Bootstrap.java
@@ -42,7 +42,6 @@ public class Bootstrap {
         logger.info("Starting car simulator ...");
         Properties joynrProperties = new Properties();
         joynrProperties.put("joynr.messaging.mqtt.brokeruris", "tcp://mqttbroker:1883");
-        joynrProperties.put(MessagingPropertyKeys.PROPERTY_MESSAGING_PRIMARYGLOBALTRANSPORT, "mqtt");
         joynrProperties.setProperty(MessagingPropertyKeys.PERSISTENCE_FILE, "car-sim-joynr.properties");
         joynrProperties.setProperty(PROPERTY_JOYNR_DOMAIN_LOCAL, "stateless_async_car_sim_local_domain");
         Module runtimeModule = Modules.combine(new CCInProcessRuntimeModule(), new HivemqMqttClientModule());